separate script file for scrolling posts into view, that can also be referenced by the contract.
